requirements.txt

Flask
gunicorn
pymongo
python-dotenv # Opcional, mas recomendado para carregar variáveis de ambiente localmente
